Add to Stage04

+++ Stage04/raster.c
@@ -58,7 +58,7 @@ void print_string(UINT8 *base, int x, int y, char str[]) {
 	
 }
 
-void print_num(UINT8 *base, int x, int y, int num) {
+void print_num(UINT8 *base, int x, int y, UINT16 num) {
 	char a, b, c, d;
 	
 	a = (num % 10) + '0';
